+++ HeatSolver/kappa_corr.f90
@@ -12,11 +12,11 @@ subroutine kappa_corr(T, kappa)
 implicit none
 
 ! Dummy
-    real, intent(in)  :: T
-    real, intent(out) :: kappa
+    real(kind=8), intent(in)  :: T
+    real(kind=8), intent(out) :: kappa
 ! Local
-    real :: a
-    real :: b  
+    real(kind=8) :: a
+    real(kind=8) :: b  
     
     a = 0.36 
     b = 0.00056
